Added the 6 Python modules
Here I've added the 6 modules of Python code! 4 of them are utilities. 2 are main programs. One of the mains is for use reading and working on CSV files in which each line is a string of text, such as titles of books. The other main is for use on a single string of text.  I believe in the latter, there is currently a string hard-coded in (which you could of course change).